requirements.txt
astroid==2.12.9
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.5.1
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mysql-connector-python==8.0.30
platformdirs==2.5.2
protobuf==3.20.1
pylint==2.15.2
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
